Incremental: make sure container name appears in /dev
We need to send a "change" event just like we do when
creating an array.

This reverts commit 382afe49b10cf3e5a4764cee74649d1cd8c91813

The problem is that we need udev to create the file in /dev
for us.
It might be unnecessary for udev to consider assembling things
in this array, but it shouldn't cause a problem.  If it did that
would be a different bug which we probably need locking to fix.

Or maybe udev shouldn't trigger a "-I" for containers appearing.

Signed-off-by: NeilBrown <neilb@suse.de>